+++ projects/triangular-engine/jolt/jolt-physics.module.ts
@@ -4,6 +4,8 @@ import { JoltPhysicsComponent } from './jolt-physics/jolt-physics.component';
 import { JoltRigidBodyComponent } from './jolt-rigid-body/jolt-rigid-body.component';
 import { JoltSphereShapeComponent } from './jolt-shapes/jolt-sphere-shape.component';
 import { JoltBoxShapeComponent } from './jolt-shapes/jolt-box-shape.component';
+import { JoltCylinderShapeComponent } from './jolt-shapes/jolt-cylinder-shape.component';
+import { JoltCapsuleShapeComponent } from './jolt-shapes/jolt-capsule-shape.component';
 import { JoltHullShapeComponent } from './jolt-shapes/jolt-hull-shape.component';
 import { JoltMeshShapeComponent } from './jolt-shapes/jolt-mesh-shape.component';
 import { JoltHeightFieldShapeComponent } from './jolt-shapes/jolt-height-field-shape.component';
@@ -16,6 +18,8 @@ const importExport = [
   JoltRigidBodyComponent,
   JoltSphereShapeComponent,
   JoltBoxShapeComponent,
+  JoltCylinderShapeComponent,
+  JoltCapsuleShapeComponent,
   JoltHullShapeComponent,
   JoltMeshShapeComponent,
   JoltHeightFieldShapeComponent,