+++ projects/triangular-engine/jolt/jolt-shapes/index.ts
@@ -1,5 +1,7 @@
 export * from './jolt-shape.component';
 export * from './jolt-box-shape.component';
+export * from './jolt-cylinder-shape.component';
+export * from './jolt-capsule-shape.component';
 export * from './jolt-sphere-shape.component';
 export * from './jolt-hull-shape.component';
 export * from './jolt-mesh-shape.component';